pyproject.toml
[tool.poetry]
name = "dbt-doctools"
version = "0.1.0"
description = "Programmatic manipulations for your dbt docs"
authors = ["Michael Cannamela"]
license = "MIT"
readme = "README.md"
packages = [{include = "dbt_doctools"}]

[tool.poetry.dependencies]
python = ">=3.8"
dbt-core = ">=1.0"
typer = "^0.4.1"
PyYAML = "^6.0"
networkx = "^2.8.4"
psycopg2-binary = "^2.9.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.2"
hypothesis = "^6.48.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
